Revert "Split up the routers (#9)"
This reverts commit 709a298800464967d8d6a153399aa6bc3249f800.
