--- api/dependencies.py
@@ -1,5 +1,0 @@
-from sqlmodel import Session, engine
-
-def get_session():
-    with Session(engine) as session:
-        yield session
